+++ src/command_merge_changes.cpp
@@ -20,6 +20,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 */
 
+#include <algorithm>
 #include <iostream>
 
 #include <boost/program_options.hpp>